Move saved questions virtual DB id to own file
