+++ frontend/src/metabase/lib/constants.js
@@ -12,5 +12,3 @@ export const NULL_DISPLAY_VALUE = t`(empty)`;
 // Must be a unique string which can be converted to a number since
 // crossfilter converts strings to numbers when grouping starts with numeric data
 export const TOTAL_ORDINAL_VALUE = "Infinity";
-
-export const SAVED_QUESTIONS_VIRTUAL_DB_ID = -1337;
