+++ frontend/src/metabase/lib/saved-questions.js
@@ -1,0 +1,1 @@
+export const SAVED_QUESTIONS_VIRTUAL_DB_ID = -1337;
