+++ frontend/src/metabase/lib/urls.js
@@ -1,6 +1,6 @@
 import slugg from "slugg";
 import { serializeCardForUrl } from "metabase/lib/card";
-import { SAVED_QUESTIONS_VIRTUAL_DB_ID } from "metabase/lib/constants";
+import { SAVED_QUESTIONS_VIRTUAL_DB_ID } from "metabase/lib/saved-questions";
 import MetabaseSettings from "metabase/lib/settings";
 import Question from "metabase-lib/lib/Question";
