+++ frontend/src/metabase/query_builder/components/saved-question-picker/SavedQuestionList.jsx
@@ -5,7 +5,7 @@ import { Box } from "grid-styled";
 
 import { PLUGIN_MODERATION } from "metabase/plugins";
 import Schemas from "metabase/entities/schemas";
-import { SAVED_QUESTIONS_VIRTUAL_DB_ID } from "metabase/lib/constants";
+import { SAVED_QUESTIONS_VIRTUAL_DB_ID } from "metabase/lib/saved-questions";
 import EmptyState from "metabase/components/EmptyState";
 import { generateSchemaId } from "metabase/schema";
